+++ components/Admin/Products/Data.js
@@ -1,0 +1,94 @@
+
+import Product1 from '../../../static/Product1.png'
+export const data = [
+    {
+        id: "1000",
+        image: Product1,
+        code: "f230fh0g3",
+        name: "Bamboo Watch",
+        description: "Product Description",
+        price: 65,
+        category: "Accessories",
+        quantity: 24,
+        inventoryStatus: "INSTOCK",
+        rating: 5,
+    },
+    {
+        id: "1001",
+        image: Product1,
+        code: "nvklal433",
+        name: "Black Watch",
+        description: "Product Description",
+        image: "black-watch.jpg",
+        price: 72,
+        category: "Accessories",
+        quantity: 61,
+        inventoryStatus: "OUTOFSTOCK",
+        rating: 4,
+    },
+    {
+        id: "1001",
+        image: Product1,
+        code: "nvklal433",
+        name: "Black Watch",
+        description: "Product Description",
+        image: "black-watch.jpg",
+        price: 72,
+        category: "Accessories",
+        quantity: 61,
+        inventoryStatus: "OUTOFSTOCK",
+        rating: 4,
+    },
+    {
+        id: "1001",
+        image: Product1,
+        code: "nvklal433",
+        name: "Black Watch",
+        description: "Product Description",
+        image: "black-watch.jpg",
+        price: 72,
+        category: "Accessories",
+        quantity: 61,
+        inventoryStatus: "OUTOFSTOCK",
+        rating: 4,
+    },
+    {
+        id: "1001",
+        image: Product1,
+        code: "nvklal433",
+        name: "Black Watch",
+        description: "Product Description",
+        image: "black-watch.jpg",
+        price: 72,
+        category: "Accessories",
+        quantity: 61,
+        inventoryStatus: "OUTOFSTOCK",
+        rating: 4,
+    },
+    {
+        id: "1001",
+        image: Product1,
+        code: "nvklal433",
+        name: "Black Watch",
+        description: "Product Description",
+        image: "black-watch.jpg",
+        price: 72,
+        category: "Accessories",
+        quantity: 61,
+        inventoryStatus: "OUTOFSTOCK",
+        rating: 4,
+    },
+    {
+        id: "1001",
+        image: Product1,
+        code: "nvklal433",
+        name: "Black Watch",
+        description: "Product Description",
+        image: "black-watch.jpg",
+        price: 72,
+        category: "Accessories",
+        quantity: 61,
+        inventoryStatus: "OUTOFSTOCK",
+        rating: 4,
+    }
+]
